# Let shuffle always use get_next_message to support holey outboxes