@tillrohrmann 2024-06-19
This PR needs to be backported to `release-1.0` and ideally also `release-0.9` to verify with the user that it fixes the problem.