@AhmedSoliman 2024-06-19
We might need to measure the performance impact of this change.