@tillrohrmann 2024-06-19
I can run both versions (before and after) with the parallel throughput benchmark to assess its effect on performance. I'll report the numbers once I have them.