@tillrohrmann 2024-06-19
Running `cargo bench --bench throughput_parallel`:
```
This PR:
throughput/parallel     time:   [650.60 ms 664.77 ms 680.18 ms]
                        thrpt:  [5.8808 Kelem/s 6.0171 Kelem/s 6.1481 Kelem/s]
main:
throughput/parallel     time:   [643.00 ms 656.26 ms 670.83 ms]
                        thrpt:  [5.9627 Kelem/s 6.0951 Kelem/s 6.2208 Kelem/s]
```

Running `cargo bench --bench throughput_sequential`:
```
This PR:
throughput/sequential   time:   [639.37 µs 645.52 µs 652.83 µs]
                        thrpt:  [1.5318 Kelem/s 1.5491 Kelem/s 1.5640 Kelem/s]
main:
throughput/sequential   time:   [625.67 µs 628.62 µs 632.37 µs]
                        thrpt:  [1.5813 Kelem/s 1.5908 Kelem/s 1.5983 Kelem/s]
```

Both benchmarks were run on my local M1Pro (note there was a bit of variance). As one can see, there is slight degradation in performance. I think the degradation is not severe enough for not shipping the hotfix. Once we have #1641 implemented, this regression should be gone.